services/telegram_bot/requirements.txt
python-telegram-bot[ext]==20.6
aio_pika
sqlalchemy
aiomysql
python-dotenv
python-json-logger
prometheus-client
fastapi
uvicorn[standard]